+++ src/Products/PloneMeeting/safe_utils.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
 # flake8: noqa
 
-from DateTime import DateTime
 from datetime import datetime
+from DateTime import DateTime
 from imio.helpers.cache import get_current_user_id
 from imio.helpers.cache import get_plone_groups_for_user
 from imio.helpers.date import wordizeDate
+from imio.helpers.workflow import update_role_mappings_for
 from imio.history.utils import getLastWFAction
 from Products.PloneMeeting.browser.views import is_all_count
 from Products.PloneMeeting.browser.views import print_votes