+++ src/Products/PloneMeeting/tests/testWFAdaptations.py
@@ -77,6 +77,7 @@ class testWFAdaptations(PloneMeetingTestCase):
                           'hide_decisions_when_under_writing_check_returned_to_proposing_group',
                           'item_validation_no_validate_shortcuts',
                           'item_validation_shortcuts',
+                          'itemdecided',
                           'mark_not_applicable',
                           MEETING_REMOVE_MOG_WFA,
                           'meetingmanager_correct_closed_meeting',
@@ -545,8 +546,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         self._activate_wfas(('postpone_next_meeting', ))
 
         meeting = self.create('Meeting')
-        item = self.create('MeetingItem')
-        item.setDecision('<p>Decision</p>')
+        item = self.create('MeetingItem', decision=self.decisionText)
         self.presentItem(item)
         self.decideMeeting(meeting)
         self.do(item, 'postpone_next_meeting')
@@ -571,20 +571,43 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.do(item, 'backToItemPublished')
         self.failIf(cfg.validate_workflowAdaptations(()))
 
+    def test_pm_Validate_workflowAdaptations_removed_itemdecided(self):
+        """Test MeetingConfig.validate_workflowAdaptations that manage removal
+           of wfAdaptations 'itemdecided' that is not possible if
+           some items are 'itemdecided'."""
+        # ease override by subproducts
+        cfg = self.meetingConfig
+        if not self._check_wfa_available(['itemdecided']):
+            return
+
+        self.changeUser('pmManager')
+        self._activate_wfas(('itemdecided', ))
+
+        meeting = self.create('Meeting')
+        item = self.create('MeetingItem', decision=self.decisionText)
+        self.presentItem(item)
+        self.decideMeeting(meeting)
+        self.do(item, 'itemdecide')
+        self.assertEqual(item.query_state(), 'itemdecided')
+        self.failIf(cfg.validate_workflowAdaptations(('itemdecided', )))
+        self.failUnless(cfg.validate_workflowAdaptations(()))
+        # make wfAdaptation selectable
+        self.do(item, 'backToItemPublished')
+        self.failIf(cfg.validate_workflowAdaptations(()))
+
     def _validate_item_decision_state_removed(self, wf_adaptation_name, item_state, item_transition):
         """Helper method checking that removing an item decision state is not
            possible if some items still in this state."""
         # ease override by subproducts
         cfg = self.meetingConfig
-        if wf_adaptation_name not in cfg.listWorkflowAdaptations():
+        if not self._check_wfa_available([wf_adaptation_name]):
             return
 
         self.changeUser('pmManager')
         self._activate_wfas((wf_adaptation_name, ))
 
         meeting = self.create('Meeting')
-        item = self.create('MeetingItem')
-        item.setDecision('<p>Decision</p>')
+        item = self.create('MeetingItem', decision=self.decisionText)
         self.presentItem(item)
         self.decideMeeting(meeting)
         self.do(item, item_transition)
@@ -670,7 +693,7 @@ class testWFAdaptations(PloneMeetingTestCase):
             item_transition='remove')
 
         cfg = self.meetingConfig
-        if 'removed_and_duplicated' in cfg.listWorkflowAdaptations():
+        if not self._check_wfa_available(['removed_and_duplicated']):
             # possible to switch from one to the other
             self.failIf(cfg.validate_workflowAdaptations(('removed_and_duplicated', )))
             cfg.setWorkflowAdaptations(('removed_and_duplicated', ))
@@ -734,32 +757,32 @@ class testWFAdaptations(PloneMeetingTestCase):
         # ease override by subproducts
         cfg = self.meetingConfig
         # accepted_out_of_meeting
-        if 'accepted_out_of_meeting' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['accepted_out_of_meeting']):
             _check(wfa_name='accepted_out_of_meeting',
                    transition='accept_out_of_meeting',
                    back_transition='backToValidatedFromAcceptedOutOfMeeting')
-        if 'accepted_out_of_meeting_and_duplicated' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['accepted_out_of_meeting_and_duplicated']):
             _check(wfa_name='accepted_out_of_meeting_and_duplicated',
                    transition='accept_out_of_meeting',
                    back_transition='backToValidatedFromAcceptedOutOfMeeting',
                    review_state='accepted_out_of_meeting')
         # accepted_out_of_meeting_emergency
-        if 'accepted_out_of_meeting_emergency' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['accepted_out_of_meeting_emergency']):
             _check(wfa_name='accepted_out_of_meeting_emergency',
                    transition='accept_out_of_meeting_emergency',
                    back_transition='backToValidatedFromAcceptedOutOfMeetingEmergency',
                    review_state='accepted_out_of_meeting_emergency')
-        if 'accepted_out_of_meeting_emergency_and_duplicated' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['accepted_out_of_meeting_emergency_and_duplicated']):
             _check(wfa_name='accepted_out_of_meeting_emergency_and_duplicated',
                    transition='accept_out_of_meeting_emergency',
                    back_transition='backToValidatedFromAcceptedOutOfMeetingEmergency',
                    review_state='accepted_out_of_meeting_emergency')
         # transfered
-        if 'transfered' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['transfered']):
             _check(wfa_name='transfered',
                    transition='transfer',
                    back_transition='backToValidatedFromTransfered')
-        if 'transfered_and_duplicated' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['transfered_and_duplicated']):
             _check(wfa_name='transfered_and_duplicated',
                    transition='transfer',
                    back_transition='backToValidatedFromTransfered',
@@ -828,9 +851,9 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.do(item, 'return_to_proposing_group')
         self.assertEqual(item.query_state(), 'returned_to_proposing_group')
         self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group', )))
-        if 'return_to_proposing_group_with_last_validation' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group_with_last_validation']):
             self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group_with_last_validation',)))
-        if 'return_to_proposing_group_with_all_validations' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group_with_all_validations']):
             self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group_with_all_validations',)))
         msg_removed_error = translate(
             'wa_removed_found_elements_error',
@@ -871,9 +894,9 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.assertEqual(item.query_state(), 'returned_to_proposing_group')
         self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group_with_last_validation',)))
         returned_to_proposing_group_last_state = 'returned_to_proposing_group_' + self._stateMappingFor('proposed')
-        if 'return_to_proposing_group' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group']):
             self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group',)))
-        if 'return_to_proposing_group_with_all_validations' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group_with_all_validations']):
             self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group_with_all_validations',)))
         self.do(item, 'goTo_' + returned_to_proposing_group_last_state)
         self.assertEqual(item.query_state(), returned_to_proposing_group_last_state)
@@ -891,11 +914,11 @@ class testWFAdaptations(PloneMeetingTestCase):
                     context=self.request)},
             context=self.request)
         self.assertEqual(cfg.validate_workflowAdaptations(()), msg_removed_error)
-        if 'return_to_proposing_group' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group']):
             self.assertEqual(
                 cfg.validate_workflowAdaptations(('return_to_proposing_group',)),
                 msg_removed_error)
-        if 'return_to_proposing_group_with_all_validations' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group_with_all_validations']):
             self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group_with_all_validations',)))
         # make wfAdaptation unselectable
         self.do(item, 'backTo_itemfrozen_from_returned_to_proposing_group')
@@ -927,21 +950,21 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group_with_all_validations',)))
         returned_to_proposing_group_last_state = 'returned_to_proposing_group_' + self._stateMappingFor('proposed')
 
-        if 'return_to_proposing_group' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group']):
             self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group', )))
 
-        if 'return_to_proposing_group_with_last_validation' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group_with_last_validation']):
             self.failIf(cfg.validate_workflowAdaptations(('return_to_proposing_group_with_last_validation',)))
         self._process_transition_for_correcting_item(item, True)
         self.assertEqual(item.query_state(), returned_to_proposing_group_last_state)
         self.assertEqual(
             cfg.validate_workflowAdaptations(()),
             return_to_proposing_group_removed_error)
-        if 'return_to_proposing_group' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group']):
             self.assertEqual(
                 cfg.validate_workflowAdaptations(('return_to_proposing_group', )),
                 return_to_proposing_group_removed_error)
-        if 'return_to_proposing_group_with_last_validation' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['return_to_proposing_group_with_last_validation']):
             self.assertEqual(
                 cfg.validate_workflowAdaptations(('return_to_proposing_group_with_last_validation',)),
                 return_to_proposing_group_removed_error)
@@ -2010,8 +2033,7 @@ class testWFAdaptations(PloneMeetingTestCase):
 
         self.changeUser('pmManager')
         meeting = self.create('Meeting')
-        item = self.create('MeetingItem')
-        item.setDecision('<p>Decision adapted by pmManager</p>')
+        item = self.create('MeetingItem', decision=self.decisionText)
         self.presentItem(item)
         self.decideMeeting(meeting)
 
@@ -2730,8 +2752,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         # test it
         self.changeUser('pmManager')
         meeting = self.create('Meeting')
-        item = self.create('MeetingItem')
-        item.setDecision('<p>A decision</p>')
+        item = self.create('MeetingItem', decision=self.decisionText)
         self.presentItem(item)
         self.decideMeeting(meeting)
         self.do(item, 'postpone_next_meeting')
@@ -2750,7 +2771,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         # test with only 'postpone_next_meeting' then when using
         # 'postpone_next_meeting' and 'no_publication' togheter if available
         set_of_wfAdaptations = [('postpone_next_meeting', )]
-        if 'no_publication' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['no_publication']):
             set_of_wfAdaptations.append(('no_publication', 'postpone_next_meeting'))
         for wfAdaptations in set_of_wfAdaptations:
             # activate the wfAdaptations and check
@@ -2809,8 +2830,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         notify(ObjectEditedEvent(cfg))
 
         self.changeUser('pmCreator1')
-        item = self.create('MeetingItem')
-        item.setDecision('<p>Decision</p>')
+        item = self.create('MeetingItem', decision=self.decisionText)
         item.setOptionalAdvisers((self.vendors_uid,
                                   '{0}__rowid__unique_id_123'.format(self.developers_uid),
                                   '{0}__rowid__unique_id_456'.format(org2_uid),
@@ -2868,8 +2888,7 @@ class testWFAdaptations(PloneMeetingTestCase):
             self._activate_wfas(('postpone_next_meeting', ), keep_existing=True)
             notify(ObjectEditedEvent(cfg))
 
-        item = self.create('MeetingItem')
-        item.setDecision('<p>Decision</p>')
+        item = self.create('MeetingItem', decision=self.decisionText)
         meeting = self.create('Meeting')
         self.presentItem(item)
         self.decideMeeting(meeting)
@@ -2918,8 +2937,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         # test it
         self.changeUser('pmManager')
         meeting = self.create('Meeting')
-        item = self.create('MeetingItem')
-        item.setDecision('<p>A decision</p>')
+        item = self.create('MeetingItem', decision=self.decisionText)
         self.presentItem(item)
         self.decideMeeting(meeting)
         self.do(item, item_transition)
@@ -3133,7 +3151,6 @@ class testWFAdaptations(PloneMeetingTestCase):
     def test_pm_WFA_accepted_out_of_meeting(self):
         '''Test the workflowAdaptation 'accepted_out_of_meeting'.'''
         # ease override by subproducts
-        cfg = self.meetingConfig
         if not self._check_wfa_available(['accepted_out_of_meeting']):
             return
         self.changeUser('pmManager')
@@ -3144,7 +3161,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         # enables it as well as in this WFA, the Review portal content permission
         # is given to reviewers on state 'validated'
         wfas = ('accepted_out_of_meeting', )
-        if 'reviewers_take_back_validated_item' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['reviewers_take_back_validated_item']):
             wfas = wfas + ('reviewers_take_back_validated_item', )
         self._activate_wfas(wfas)
         self._accepted_out_of_meeting_active()
@@ -3184,7 +3201,7 @@ class testWFAdaptations(PloneMeetingTestCase):
 
         # test 'accepted_out_of_meeting_and_duplicated' if available
         cfg = self.meetingConfig
-        if 'accepted_out_of_meeting_and_duplicated' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['accepted_out_of_meeting_and_duplicated']):
             wfas = list(cfg.getWorkflowAdaptations())
             wfas.remove('accepted_out_of_meeting')
             wfas.append('accepted_out_of_meeting_and_duplicated')
@@ -3199,7 +3216,6 @@ class testWFAdaptations(PloneMeetingTestCase):
     def test_pm_WFA_accepted_out_of_meeting_emergency(self):
         '''Test the workflowAdaptation 'accepted_out_of_meeting_emergency'.'''
         # ease override by subproducts
-        cfg = self.meetingConfig
         if not self._check_wfa_available(['accepted_out_of_meeting_emergency']):
             return
         self.changeUser('pmManager')
@@ -3210,7 +3226,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         # enables it as well as in this WFA, the Review portal content permission
         # is given to reviewers on state 'validated'
         wfas = ('accepted_out_of_meeting_emergency', )
-        if 'reviewers_take_back_validated_item' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['reviewers_take_back_validated_item']):
             wfas = wfas + ('reviewers_take_back_validated_item', )
         self._activate_wfas(wfas)
         self._accepted_out_of_meeting_emergency_active()
@@ -3232,6 +3248,9 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.validateItem(item)
         # not available until MeetingItem.isAcceptableOutOfMeeting is True
         self.assertFalse('accept_out_of_meeting_emergency' in self.transitions(item))
+        item.setIsAcceptableOutOfMeeting(True)
+        # not available until emergency is 'emergency_accepted'
+        self.assertFalse('accept_out_of_meeting_emergency' in self.transitions(item))
         item.setEmergency('emergency_accepted')
         self.assertTrue('accept_out_of_meeting_emergency' in self.transitions(item))
         # in case 'reviewers_take_back_validated_item' is available
@@ -3250,7 +3269,7 @@ class testWFAdaptations(PloneMeetingTestCase):
 
         # test 'accepted_out_of_meeting_emergency_and_duplicated' if available
         cfg = self.meetingConfig
-        if 'accepted_out_of_meeting_emergency_and_duplicated' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['accepted_out_of_meeting_emergency_and_duplicated']):
             wfas = list(cfg.getWorkflowAdaptations())
             wfas.remove('accepted_out_of_meeting_emergency')
             wfas.append('accepted_out_of_meeting_emergency_and_duplicated')
@@ -3261,11 +3280,12 @@ class testWFAdaptations(PloneMeetingTestCase):
             self.assertEqual(duplicated_item.query_state(), 'validated')
             # duplicated_item emergency is no more asked
             self.assertEqual(duplicated_item.getEmergency(), 'no_emergency')
+            # duplicated_item is not more isAcceptableOutOfMeeting
+            self.assertFalse(duplicated_item.getIsAcceptableOutOfMeeting())
 
     def test_pm_WFA_transfered(self):
         '''Test the workflowAdaptation 'transfered'.'''
         # ease override by subproducts
-        cfg = self.meetingConfig
         if not self._check_wfa_available(['transfered']):
             return
         self.changeUser('pmManager')
@@ -3276,7 +3296,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         # enables it as well as in this WFA, the Review portal content permission
         # is given to reviewers on state 'validated'
         wfas = ('transfered', )
-        if 'reviewers_take_back_validated_item' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['reviewers_take_back_validated_item']):
             wfas = wfas + ('reviewers_take_back_validated_item', )
         self._activate_wfas(wfas)
         self._transfered_active()
@@ -3316,7 +3336,7 @@ class testWFAdaptations(PloneMeetingTestCase):
 
         # test 'transfered_and_duplicated' if available
         cfg = self.meetingConfig
-        if 'transfered_and_duplicated' in cfg.listWorkflowAdaptations():
+        if self._check_wfa_available(['transfered_and_duplicated']):
             wfas = list(cfg.getWorkflowAdaptations())
             wfas.remove('transfered')
             wfas.append('transfered_and_duplicated')
@@ -3367,6 +3387,113 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.assertTrue(meeting.wfConditions().mayCorrect())
         self.changeUser('siteadmin')
         self.assertTrue(meeting.wfConditions().mayCorrect())
+
+    def test_pm_ItemDecided(self):
+        '''Test the workflowAdaptation 'itemdecided'.'''
+        # ease override by subproducts
+        if not self._check_wfa_available(['itemdecided']):
+            return
+        self.changeUser('pmManager')
+        # check while the wfAdaptation is not activated
+        self._itemdecided_inactive()
+        self._activate_wfas(('itemdecided', ))
+        self._itemdecided_active()
+
+    def _itemdecided_inactive(self):
+        '''Tests while 'itemdecided' wfAdaptation is inactive.'''
+        self.changeUser('pmManager')
+        item = self.create('MeetingItem')
+        meeting = self.create('Meeting')
+        self.presentItem(item)
+        self.decideMeeting(meeting)
+        self.assertEqual(item.query_state(), 'itempublished')
+        self.assertEqual(self.transitions(item), ['backToItemFrozen'])
+
+    def _itemdecided_active(self):
+        '''Tests while 'transfered' wfAdaptation is active.'''
+        self.changeUser('pmManager')
+        item = self.create('MeetingItem', decision=self.decisionText)
+        meeting = self.create('Meeting')
+        self.presentItem(item)
+        self.decideMeeting(meeting)
+        self.assertEqual(self.transitions(item), ['backToItemFrozen', 'itemdecide'])
+        self.changeUser('pmCreator1')
+        self.assertFalse(self.transitions(item))
+        self.changeUser('pmManager')
+        self.do(item, 'itemdecide')
+        self.assertEqual(item.query_state(), 'itemdecided')
+        # now item may be decided
+        self.assertEqual(self.transitions(item), ['accept', 'backToItemPublished'])
+
+    def test_pm_ItemDecidedWithNoFreezeNoPublicationNoDecide(self):
+        """Test with "itemdecided" but "no_freeze/no_decide/no_publication"."""
+        cfg = self.meetingConfig
+        if not self._check_wfa_available(
+                ['itemdecided', 'no_publication', 'no_freeze', 'no_decide']):
+            return
+        self._activate_wfas(())
+        # states are there by default
+        itemWFStates = cfg.getItemWorkflow(True).states
+        self.assertTrue('itemfrozen' in itemWFStates)
+        self.assertTrue('itempublished' in itemWFStates)
+        self.assertFalse('itemdecided' in itemWFStates)
+        meetingWFStates = cfg.getMeetingWorkflow(True).states
+        self.assertTrue('frozen' in meetingWFStates)
+        self.assertTrue('published' in meetingWFStates)
+        self.assertTrue('decided' in meetingWFStates)
+        # enable "itemdecided" and some others, this will also test that WFAs
+        # are applied without problem
+        self._activate_wfas(('itemdecided', 'no_publication'))
+        itemWFStates = cfg.getItemWorkflow(True).states
+        self.assertTrue('itemfrozen' in itemWFStates)
+        self.assertFalse('itempublished' in itemWFStates)
+        self.assertTrue('itemdecided' in itemWFStates)
+        meetingWFStates = cfg.getMeetingWorkflow(True).states
+        self.assertTrue('frozen' in meetingWFStates)
+        self.assertFalse('published' in meetingWFStates)
+        self.assertTrue('decided' in meetingWFStates)
+        # remove as much as possible
+        self._activate_wfas(('itemdecided', 'no_publication', 'no_freeze', 'no_decide'))
+        itemWFStates = cfg.getItemWorkflow(True).states
+        self.assertFalse('itemfrozen' in itemWFStates)
+        self.assertFalse('itempublished' in itemWFStates)
+        self.assertTrue('itemdecided' in itemWFStates)
+        meetingWFStates = cfg.getMeetingWorkflow(True).states
+        self.assertFalse('frozen' in meetingWFStates)
+        self.assertFalse('published' in meetingWFStates)
+        self.assertFalse('decided' in meetingWFStates)
+
+    def test_pm_ItemDecidedWithReturnToProposingGroup(self):
+        """Test with "itemdecided" together with "return_to_proposing_group"
+           when meeting frozen, item can go back to "itemfrozen" and
+           when meeting decided, item can go back to "itemdecided"."""
+        if not self._check_wfa_available(
+                ['itemdecided', 'return_to_proposing_group']):
+            return
+        # enable auto itemdecide item when meeting decided
+        cfg = self.meetingConfig
+        actions = list(cfg.getOnMeetingTransitionItemActionToExecute())
+        actions.insert(5,
+                       {'meeting_transition': 'decide',
+                        'item_action': 'itemdecide',
+                        'tal_expression': ''})
+        cfg.setOnMeetingTransitionItemActionToExecute(actions)
+        self._activate_wfas(('itemdecided', 'return_to_proposing_group'))
+        self.changeUser('pmManager')
+        item = self.create('MeetingItem', decision=self.decisionText)
+        meeting = self.create('Meeting')
+        self.presentItem(item)
+        self.do(item, 'return_to_proposing_group')
+        self.assertEqual(self.transitions(item),
+                         ['backTo_presented_from_returned_to_proposing_group'])
+        self.freezeMeeting(meeting)
+        self.assertEqual(self.transitions(item),
+                         ['backTo_itemfrozen_from_returned_to_proposing_group'])
+        self.decideMeeting(meeting)
+        self.assertEqual(self.transitions(item),
+                         ['backTo_itemdecided_from_returned_to_proposing_group'])
+        self.do(item, 'backTo_itemdecided_from_returned_to_proposing_group')
+        self.assertEqual(item.query_state(), 'itemdecided')
 
 
 def test_suite():